+++ jdaviz/core/loaders/test_loaders.py
@@ -129,11 +129,9 @@ def test_fits_spectrum2d(deconfigged_helper):
         ldr = deconfigged_helper.loaders['file']
         ldr.filepath = uri
 
-    # Default is Image but the test switches to 2D Spectrum
-    # since this file type is not yet supported by the image loader
+    # Default format is Image, manually set to 2D Spectrum
     assert ldr.format == 'Image'
     assert ldr.importer._obj.input_has_extensions is True
-
     ldr.format = '2D Spectrum'
 
     ldr.importer()
@@ -249,6 +247,34 @@ def test_invoke_from_plugin(specviz_helper, spectrum1d, tmp_path):
     assert len(loader.format.choices) > 0
 
     loader.importer()
+
+
+@pytest.mark.parametrize('order', ([0, 1, 2], [0, 2, 1], [2, 0, 1], [2, 1, 0], [1, 2, 0]))
+def test_mult_data_types(deconfigged_helper, image_nddata_wcs, spectrum2d, spectrum1d, order):
+    datas = [image_nddata_wcs, spectrum2d, spectrum1d]
+    load_kwargs = [{'format': 'Image'},
+                   {'format': '2D Spectrum', 'auto_extract': True},
+                   {'format': '1D Spectrum'}]
+
+    for i in order:
+        deconfigged_helper.load(datas[i], **load_kwargs[i])
+
+    # NOTE: 2D Spectrum will also result in auto-extracted 1D Spectrum
+    assert len(deconfigged_helper.app.data_collection) == 4
+    assert len(deconfigged_helper.viewers) == 3
+
+
+def test_freq_wavelength_linking(deconfigged_helper, spectrum1d):
+    deconfigged_helper.load(spectrum1d, format='1D Spectrum', data_label='sp_wavelength')
+    sp1d_freq = Spectrum(spectral_axis=spectrum1d.spectral_axis.to(u.Hz, equivalencies=u.spectral()),  # noqa
+                         flux=spectrum1d.flux,
+                         uncertainty=spectrum1d.uncertainty,
+                         mask=spectrum1d.mask,
+                         meta=spectrum1d.meta)
+    deconfigged_helper.load(sp1d_freq, format='1D Spectrum', data_label='sp_frequency')
+
+    # flux <> flux, uncertainty <> uncertainty, wavelength <> freq
+    assert len(deconfigged_helper.app.data_collection.external_links) == 3
 
 
 def test_load_image_mult_sci_extension(imviz_helper):